Example format of research awards data. Data is not real.

An example set of what some awards data for a UK university may look like.

commoncalcs_df

Format

A data frame with 210 rows and 7 variables.

college

part of organisational structure, could also be faculty

hesa_category

classification of funder of research using UK HESA system

actual_start_date

date the project started, or is due to start

actual_end_date

date project ended, or is due to end

award_value

value funder awarded to university

application_value

value university asked of the funder

award_ac_yr

year award was granted by funder