Create a 'straight-line' profile of a value over its associated lifetime, split up into user-defined periods.

Can take a value associated with an item and spreads the value over the item's lifetime. One example would be profiling income/expenditure using the total value of a five year contract.

profile_over_time(
  df,
  begin_col,
  end_col,
  period_start,
  period_end,
  period_breaks,
  value_col,
  ...
)

Arguments

df
Dataframe holding the data you would like to perform the calculation for.

begin_col
Column containing information about the begin date for the item you're interested in. Must be in date format. Will be renamed 'start_date' during processing.

end_col
Column ontaining the information about the end date of the item you're interested in. Must be in date format. Will be renamed 'end_date' during processing.

period_start
The beginning of the period you would like to profile over. Enter as "yyyy-mm-dd".

period_end
The end of the period you would like to profile over. Enter as "yyyy-mm-dd".

period_breaks
Defines the length of each period. Can be "day", "week", "month", "quarter" or "year".

value_col
The value you'd like to profile. Must be numeric.

...
Columns containing information about the item the value is attached to.

Value

A dataframe containing the data profiled over the specified periods.